refactor:(backend) ♻️ change folder name
