+++ backend/src/modules/auth/dto/sign-in-response.dto.ts
