+++ backend/src/modules/auth/entities/auth.entity.ts
