+++ backend/src/modules/auth/strategies/jwt.strategy.ts
