+++ backend/src/modules/user/entities/enums/seniority.enum.ts
