+++ backend/src/modules/user/entities/user.entity.ts
